pyproject.toml
[project]
name = "openrouter-discord-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.7.1",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "tiktoken>=0.12.0",
]

[tool.mypy]
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[tool.ruff]
line-length = 200
target-version = "py311"
ignore = ["RUF001", "RUF002"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "RUF"]

[dependency-groups]
dev = ["mypy>=1.20.0", "ruff>=0.15.9"]
# I - isort (сортировка импортов, как simple-import-sort)
# B - flake8-bugbear (поиск багов)
# UP - pyupgrade (современный синтаксис)
